feat(sweet-alert): show message with sweet alert

+++ src/app/components/list-employee/list-employee.component.ts
@@ -5,6 +5,7 @@ import { HttpClient } from '@angular/common/http';
 import { EmployeeService } from '../../services/employee.service';
 import { tap } from 'rxjs';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-employee',
@@ -44,14 +45,31 @@ export class ListEmployeeComponent implements OnInit {
   }
 
   deleteEmployee(id: number): void {
-    this._employeeService
-      .deleteEmployee(id)
-      .pipe(
-        tap((empleados) => {
-          this.listOfEmployee = empleados;
-        })
-      )
-      .subscribe();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._employeeService
+          .deleteEmployee(id)
+          .pipe(
+            tap((empleados) => {
+              this.listOfEmployee = empleados;
+              Swal.fire({
+                title: 'Deleted!',
+                text: 'Employee has been deleted.',
+                icon: 'success',
+              });
+            })
+          )
+          .subscribe();
+      }
+    });
   }
 
   editEmployee(employee: IEmployee): void {
